human pose detection
MoveNet is a deep learning model used for human pose detection, which involves identifying and tracking body poses in images or videos. By estimating key points on a person's body, such as joints and landmarks, MoveNet can detect and track the positions of various body parts. This information is valuable for applications like action recognition, fitness tracking, augmented reality, and human-computer interaction. MoveNet is known for its accuracy and speed, making it a popular choice for real-time pose estimation tasks.